// web/src/api/models/PlexMediaItemData.ts
/* tslint:disable */
/* eslint-disable */
/**
 * mindscape
 * This is the swagger page for the Project mindscape generated with Egg-go. use this to test your database connection
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PlexMediaItemData
 */
export interface PlexMediaItemData {
    /**
     * 
     * @type {number}
     * @memberof PlexMediaItemData
     */
    addedAt?: number;
    /**
     * 
     * @type {string}
     * @memberof PlexMediaItemData
     */
    parentTitle?: string;
    /**
     * 
     * @type {string}
     * @memberof PlexMediaItemData
     */
    thumb?: string;
    /**
     * 
     * @type {string}
     * @memberof PlexMediaItemData
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof PlexMediaItemData
     */
    type?: string;
    /**
     * 
     * @type {number}
     * @memberof PlexMediaItemData
     */
    year?: number;
}

/**
 * Check if a given object implements the PlexMediaItemData interface.
 */
export function instanceOfPlexMediaItemData(value: object): value is PlexMediaItemData {
    return true;
}

export function PlexMediaItemDataFromJSON(json: any): PlexMediaItemData {
    return PlexMediaItemDataFromJSONTyped(json, false);
}

export function PlexMediaItemDataFromJSONTyped(json: any, ignoreDiscriminator: boolean): PlexMediaItemData {
    if (json == null) {
        return json;
    }
    return {
        
        'addedAt': json['added_at'] == null ? undefined : json['added_at'],
        'parentTitle': json['parent_title'] == null ? undefined : json['parent_title'],
        'thumb': json['thumb'] == null ? undefined : json['thumb'],
        'title': json['title'] == null ? undefined : json['title'],
        'type': json['type'] == null ? undefined : json['type'],
        'year': json['year'] == null ? undefined : json['year'],
    };
}

export function PlexMediaItemDataToJSON(json: any): PlexMediaItemData {
    return PlexMediaItemDataToJSONTyped(json, false);
}

export function PlexMediaItemDataToJSONTyped(value?: PlexMediaItemData | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'added_at': value['addedAt'],
        'parent_title': value['parentTitle'],
        'thumb': value['thumb'],
        'title': value['title'],
        'type': value['type'],
        'year': value['year'],
    };
}